:art: get_data, etc is now mocked
