+++ api/namespace/applications/helpers/helpers.py
@@ -23,6 +23,7 @@ class ApplicationHelpers:
         fund_round = get_round(fund_id, round_id)
         if fund and fund_round:
             fund_round_forms = Config.FUND_ROUND_FORMS
+            print("DDDDDDDD", fund_round_forms)
             forms = fund_round_forms.get(":".join([fund_id, round_id]))
             if not forms:
                 raise Exception(
