+++ external_services/data.py
@@ -1,89 +1,17 @@
 import json
 import os
 import urllib.parse
-import shutil
 from typing import List
-from flask import current_app
-
-import requests
 from config import Config
 from external_services.models.fund import Fund
 from external_services.models.round import Round
+from external_services import http_methods
 from flask import current_app
-
-
-def api_call(endpoint: str, method: str = "GET", params: dict = None):
-    if params:
-        params = {k: v for k, v in params.items() if v is not None}
-    if endpoint.startswith("http"):
-        if method:
-            if method == "POST":
-                return requests.post(endpoint, json=params)
-            elif method == "GET":
-                req = requests.PreparedRequest()
-                req.prepare_url(endpoint, params)
-                return requests.get(req.url)
-    else:
-        return local_api_call(endpoint, params, method)
-
-
-def get_data(endpoint: str, params: dict = None):
-    if params:
-        params = {k: v for k, v in params.items() if v is not None}
-    if endpoint.startswith("http"):
-        req = requests.PreparedRequest()
-        req.prepare_url(endpoint, params)
-        current_app.logger.info(f"HTTP request made to {endpoint}")
-        response = requests.get(req.url)
-        if response.status_code == 200:
-            return response.json()
-    else:
-        current_app.logger.info(f"Local api call made to {endpoint}")
-        return local_api_call(endpoint, params, "get")
-
-
-def post_data(endpoint: str, params: dict = None):
-    if params:
-        params = {k: v for k, v in params.items() if v is not None}
-    if endpoint.startswith("http"):
-        response = requests.post(endpoint, json=params)
-        if response.status_code in [200, 201]:
-            return response.json()
-    else:
-        return local_api_call(endpoint, params, "post")
-
-
-def local_api_call(endpoint: str, params: dict = None, method: str = "get"):
-    api_data_json = os.path.join(
-        Config.FLASK_ROOT,
-        "tests",
-        "api_data",
-        method.lower() + "_endpoint_data.json",
-    )
-    fp = open(api_data_json)
-    api_data = json.load(fp)
-    fp.close()
-    query_params = "_"
-    if params:
-        query_params = urllib.parse.urlencode(params)
-    if method.lower() == "post":
-        if endpoint in api_data:
-            post_dict = api_data.get(endpoint)
-            if query_params in post_dict:
-                return post_dict.get(query_params)
-            else:
-                return post_dict.get("_default")
-    else:
-        if params:
-            endpoint = f"{endpoint}?{query_params}"
-        if endpoint in api_data:
-            return api_data.get(endpoint)
 
 
 def get_funds() -> List[Fund] | None:
     endpoint = Config.FUND_STORE_API_HOST + Config.FUNDS_ENDPOINT
-    current_app.logger.info(f"Making request to:'{endpoint}'")
-    response = get_data(endpoint)
+    response = http_methods.get_data(endpoint)
     if response and len(response) > 0:
         funds = []
         for fund in response:
@@ -96,20 +24,19 @@ def get_fund(fund_id: str) -> Fund | None:
         fund_id=fund_id
     )
     current_app.logger.info(f"Request made to {endpoint}")
-    response = get_data(endpoint)
-    if response and "fund_id" in response:
-        fund = Fund.from_json(response)
-        if "rounds" in response and len(response["rounds"]) > 0:
-            for fund_round in response["rounds"]:
-                fund.add_round(Round.from_json(fund_round))
-        return fund
+    response = http_methods.get_data(endpoint)
+    print(response)
+    if response is None:
+        current_app.logger.info("Request to fund store returned None")
+    fund = Fund.from_json(response)
+    return fund
 
 
 def get_rounds(fund_id: str) -> Fund | List:
     endpoint = Config.FUND_STORE_API_HOST + Config.FUND_ROUNDS_ENDPOINT.format(
         fund_id=fund_id
     )
-    response = get_data(endpoint)
+    response = http_methods.get_data(endpoint)
     rounds = []
     if response and len(response) > 0:
         for round_data in response:
@@ -125,8 +52,6 @@ def get_round(fund_id: str, round_id: str) -> Round | None:
         Config.FUND_STORE_API_HOST
         + Config.FUND_ROUND_ENDPOINT.format(fund_id=fund_id, round_id=round_id)
     )
-    current_app.logger.info(f"Making request to:'{round_endpoint}'")
-    round_response = get_data(round_endpoint)
-    current_app.logger.info(f"Made request to:'{round_endpoint}', found round {round_response}")
+    round_response = http_methods.get_data(round_endpoint)
     if round_response and "id" in round_response:
         return Round.from_json(round_response)
