+++ tests/test_routes.py
@@ -285,9 +285,8 @@ def test_get_application_by_application_id(client):
     )
 
 
-def testHealthcheckRoute(flask_test_client):
-    expected_result = {"checks": [{"check_running": "OK"}]}
-
-    result = flask_test_client.get("/healthcheck")
+def testHealthcheckRoute(client):
+    expected_result = {"checks": [{"check_flask_running": "OK"}]}
+    result = client.get("/healthcheck")
     assert result.status_code == 200, "Unexpected status code"
     assert result.json == expected_result, "Unexpected json body"
